Port to Windows
The "all warnings" setting we have for gcc may be a bit overkill here,
as I end up turning off a whole bunch of them. We'll see how it works
out.

Packing works differently on MSVC than it does on non-Microsoft
compilers, so that's been adjusted. cmake submodule needs an update, and
we need to link with more stuff from Boost for some reason. I don't
think that hurts the non-Microsoft side, so I didn't make that
Windows-only.