Fix breaking the One-Definition Rule
profile.hpp (formerly config.h) defines the protocol's profile and can
radically change the definition of various types. We need it
pre-included for anything to work.

Static asserts make sure this is done correctly

This will also allow me to automatically generate and test all possible
permutations of the options.