Add documentation for an option discussed and rejected
This option, including the association IDs in the session messages, was part of my 2017 strawman proposal and made its way into the SG15 working documents for 62351-5, but has already been rejected by the DNP SATF due to its not being necessary. I've added it to the profile.hpp file, but not to the gen_profile script because I don't intend to implement it, at least for now.