Remove the AID option from the profiles
It's no longer an option: at the DNP 2018 F2F, it was decided that we
need the AID and a use-case in which the Outstation would be unable to
distinguish between two redundant Masters was presented. Those Masters
cannot be allowed to share Update Keys, so they shall have different
ascocation IDs.